requirements.txt
asgiref==3.7.2
boto3==1.34.144
botocore==1.34.144
certifi==2023.7.22
charset-normalizer==3.1.0
distlib==0.3.8
Django==4.2.2
django-cors-headers==4.3.1
django-filter==23.2
django-query-counter==0.3.1
django-storages==1.14.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.14.0
idna==3.4
jmespath==1.0.1
Markdown==3.4.3
Pillow==10.0.1
platformdirs==4.2.2
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.9.0.post0
pytz==2023.3
requests==2.31.0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.4.4
tabulate==0.9.0
urllib3==2.0.7
virtualenv==20.26.2
